test data real sus removed firstlast name for employee entity, changed some other random things nd added test data in singleton

+++ HotelReservationSystem-ejb/src/java/ejb/session/stateless/RoomSessionBeanLocal.java
@@ -26,7 +26,7 @@ public interface RoomSessionBeanLocal {
     public RoomEntity createNewRoom(RoomEntity newRoomEntity, String roomTypeName) throws RoomNameExistException, UnknownPersistenceException, InputDataValidationException, RoomTypeNotFoundException;
     public List<RoomEntity> retrieveAllRooms();
     public RoomEntity retrieveRoomByRoomId(Long roomId) throws RoomNotFoundException;
-    public RoomEntity retrieveRoomByRoomNumber(long number) throws RoomNotFoundException;
+    public RoomEntity retrieveRoomByRoomNumber(Integer number) throws RoomNotFoundException;
     public void updateRoom(RoomEntity roomEntity) throws RoomNotFoundException, UpdateRoomException, InputDataValidationException;
     public void deleteRoom(Long roomId) throws RoomNotFoundException, DeleteRoomException;
 }
